Benchmark refactoring to make them run in github actions
