+++ benchmarks/MiniExcel.Benchmarks/BenchmarkSections/XlsxAsyncBenchmark.cs
@@ -1,7 +1,7 @@
 ﻿using BenchmarkDotNet.Attributes;
 using MiniExcelLibs.Benchmarks.Utils;
 
-namespace MiniExcelLibs.Benchmarks;
+namespace MiniExcelLibs.Benchmarks.BenchmarkSections;
 
 public class XlsxAsyncBenchmark : BenchmarkBase
 {
@@ -22,7 +22,7 @@ public class XlsxAsyncBenchmark : BenchmarkBase
         using var path = AutoDeletingPath.Create();
         var value = new
         {
-            employees = Enumerable.Range(1, rowCount)
+            employees = Enumerable.Range(1, RowCount)
                 .Select(s => new 
                 {
                     name = "Jack",
