+++ benchmarks/MiniExcel.Benchmarks/Program.cs
@@ -1,13 +1,22 @@
-﻿using BenchmarkDotNet.Running;
+﻿using System.ComponentModel;
+using BenchmarkDotNet.Running;
 using MiniExcelLibs.Benchmarks;
+using MiniExcelLibs.Benchmarks.BenchmarkSections;
 
-
-#if DEBUG
-new XlsxBenchmark().Epplus_QueryFirst_Test();
-#else
-BenchmarkSwitcher
-    .FromTypes([typeof(XlsxBenchmark)])
-    .Run(args, new Config());
-#endif
-
-Console.Read();
+if (Environment.GetEnvironmentVariable("BenchmarkMode") == "Automatic")
+{
+    var section = Environment.GetEnvironmentVariable("BenchmarkSection");
+    var benchmark = section?.ToLowerInvariant().Trim() switch
+    {
+        "query" => typeof(QueryXlsxBenchmark),
+        "create" => typeof(CreateXlsxBenchmark),
+        "template" => typeof(TemplateXlsxBenchmark),
+        _ => throw new InvalidEnumArgumentException($"Benchmark section {section} does not exist")
+    };
+    
+    BenchmarkRunner.Run(benchmark, new Config(), args);
+}
+else
+    BenchmarkSwitcher
+        .FromTypes([typeof(CreateXlsxBenchmark)])
+        .Run(args, new Config());
