+++ benchmarks/MiniExcel.Benchmarks/Utils/Extensions.cs
@@ -7,6 +7,14 @@ internal static class Extensions
     internal static void Add(this Row row, params string[] values)
     {
         foreach (var value in values)
-            row.Append(new Cell { CellValue = new CellValue(value), DataType = CellValues.String });
+        {
+            var cell = new Cell
+            {
+                CellValue = new CellValue(value),
+                DataType = CellValues.String
+            };
+            
+            row.Append(cell);
+        }
     }
 }
